+++ include/lingebra.hpp
@@ -6,28 +6,28 @@
 
 
 /**
- * Vector of doubles with all necessary operators defined 
+ * Vector of floats with all necessary operators defined 
  */
 
 class DoubleVec
 {
-    std::vector<double> _vec;
+    std::vector<float> _vec;
 
 public:
     DoubleVec(): _vec() {}
 
     /* construct a zero DoubleVec of given dimension */
-    explicit DoubleVec(int dimension): _vec(std::vector<double>(dimension)) {}
+    explicit DoubleVec(int dimension): _vec(std::vector<float>(dimension)) {}
 
-    /* construct a DoubleVec from given vector of doubles */
-    explicit DoubleVec(std::vector<double> &v): _vec(v) {}
-    explicit DoubleVec(std::vector<double> &&v): _vec(std::move(v)) {}
+    /* construct a DoubleVec from given vector of floats */
+    explicit DoubleVec(std::vector<float> &v): _vec(v) {}
+    explicit DoubleVec(std::vector<float> &&v): _vec(std::move(v)) {}
 
     /* std::vector operations */
     int size() const { return _vec.size(); }
     bool empty() const { return _vec.empty(); }
-    void push_back(double num) { _vec.push_back(num); }
-    void pop_back(double num) { _vec.pop_back(); }
+    void push_back(float num) { _vec.push_back(num); }
+    void pop_back(float num) { _vec.pop_back(); }
     void reserve(int n) { _vec.reserve(n); }
 
     /* Arithmetic operators that modify sthe object */
@@ -50,7 +50,7 @@ public:
         return *this;
     }
 
-    DoubleVec& operator*=(double scalar)
+    DoubleVec& operator*=(float scalar)
     {
         for (int i = 0; i < size(); ++i) {
             _vec[i] *= scalar;
@@ -60,7 +60,7 @@ public:
 
     // having *= for dot product does not make sense
 
-    DoubleVec& operator/=(double scalar)
+    DoubleVec& operator/=(float scalar)
     {
         for (int i = 0; i < size(); ++i) {
             _vec[i] /= scalar;
@@ -73,18 +73,18 @@ public:
     friend DoubleVec operator-(DoubleVec vec);
     friend DoubleVec operator+(DoubleVec first, const DoubleVec& second);
     friend DoubleVec operator-(DoubleVec first, const DoubleVec& second);
-    friend double operator*(const DoubleVec& first, const DoubleVec& second);
+    friend float operator*(const DoubleVec& first, const DoubleVec& second);
 
-    friend DoubleVec operator*(DoubleVec vec, double scalar);
-    friend DoubleVec operator*(double scalar, DoubleVec vec);
-    friend DoubleVec operator/(DoubleVec vec, double scalar);
-    friend DoubleVec operator/(double scalar, DoubleVec vec);
+    friend DoubleVec operator*(DoubleVec vec, float scalar);
+    friend DoubleVec operator*(float scalar, DoubleVec vec);
+    friend DoubleVec operator/(DoubleVec vec, float scalar);
+    friend DoubleVec operator/(float scalar, DoubleVec vec);
 
     /* Squares the inside values */
     friend DoubleVec square_inside(DoubleVec vec);
 
     /* adds scalar to every item in vector */
-    friend DoubleVec add_scalar_to_all_items(DoubleVec vec, double d);
+    friend DoubleVec add_scalar_to_all_items(DoubleVec vec, float d);
 
     /* Relational operators */
     bool operator==(const DoubleVec& other) const { return this->_vec == other._vec; }
@@ -92,23 +92,23 @@ public:
 
     /* Indexing and iterating */
 
-    double operator[](int i) const 
+    float operator[](int i) const 
     { 
         assert(i >= 0 && i <= size() - 1);
         return _vec[i]; 
     }
 
-    double& operator[](int i) 
+    float& operator[](int i) 
     { 
         assert(i >= 0 && i <= size() - 1);
         return _vec[i]; 
     }
 
-    std::vector<double>::iterator begin() { return _vec.begin(); }
-    std::vector<double>::iterator end() { return _vec.end(); }
+    std::vector<float>::iterator begin() { return _vec.begin(); }
+    std::vector<float>::iterator end() { return _vec.end(); }
 
-    std::vector<double>::const_iterator begin() const { return _vec.begin(); }
-    std::vector<double>::const_iterator end() const { return _vec.end(); }
+    std::vector<float>::const_iterator begin() const { return _vec.begin(); }
+    std::vector<float>::const_iterator end() const { return _vec.end(); }
 
     void compareSizes(const DoubleVec& other) const;
 };
@@ -130,10 +130,10 @@ DoubleVec operator-(DoubleVec first, const DoubleVec& second)
 { return first -= second; }
 
 /* dot product */
-double operator*(const DoubleVec& first, const DoubleVec& second)
+float operator*(const DoubleVec& first, const DoubleVec& second)
 {
     first.compareSizes(second);
-    double result = 0;
+    float result = 0;
     for (int i = 0; i < first.size(); ++i) {
         result += first[i] * second[i];
     }
@@ -141,16 +141,16 @@ double operator*(const DoubleVec& first, const DoubleVec& second)
 }
 
 /* Multiplication and division by scalar */
-DoubleVec operator*(DoubleVec vec, double scalar)
+DoubleVec operator*(DoubleVec vec, float scalar)
 { return vec *= scalar; }
 
-DoubleVec operator*(double scalar, DoubleVec vec)
+DoubleVec operator*(float scalar, DoubleVec vec)
 { return vec *= scalar; }
 
-DoubleVec operator/(DoubleVec vec, double scalar)
+DoubleVec operator/(DoubleVec vec, float scalar)
 { return vec /= scalar; }
 
-DoubleVec operator/(double scalar, DoubleVec vec)
+DoubleVec operator/(float scalar, DoubleVec vec)
 { return vec /= scalar; }
 
 /* Squares the inside values */
@@ -172,7 +172,7 @@ DoubleVec sqrt_inside(DoubleVec vec)
 }
 
 /* adds scalar to every item in vector */
-DoubleVec add_scalar_to_all_items(DoubleVec vec, double d)
+DoubleVec add_scalar_to_all_items(DoubleVec vec, float d)
 {      
     for (int i = 0; i < vec.size(); ++i) {
         vec[i] += d;
@@ -206,7 +206,7 @@ void DoubleVec::compareSizes(const DoubleVec& other) const {
 
 
 /**
- * Matrix of doubles with all necessary operators defined 
+ * Matrix of floats with all necessary operators defined 
  */
 
 class DoubleMat
@@ -289,7 +289,7 @@ public:
     }
 
     /* multiplying by a scalar */
-    DoubleMat &operator*=(double scal)
+    DoubleMat &operator*=(float scal)
     {
         for (int i = 0; i < row_num(); ++i) {
             _matrix_rows[i] *= scal;
@@ -298,7 +298,7 @@ public:
     }
 
     /* dividing by a scalar */
-    DoubleMat &operator/=(double scal)
+    DoubleMat &operator/=(float scal)
     {
         for (int i = 0; i < row_num(); ++i) {
             _matrix_rows[i] /= scal;
@@ -310,7 +310,7 @@ public:
 
     friend DoubleMat square_inside(DoubleMat mat);
 
-    friend DoubleMat add_scalar_to_all_items(DoubleMat mat, double d);
+    friend DoubleMat add_scalar_to_all_items(DoubleMat mat, float d);
 
     /* adds given vector to all rows */
     DoubleMat &add_vec_to_all_rows(const DoubleVec &vec)
@@ -333,7 +333,7 @@ public:
     }
 
     /* Computes dot product of row of first matrix and col of second matrix */
-    friend double dot_row_col(const DoubleMat& first, const DoubleMat& second, int row, int col);
+    friend float dot_row_col(const DoubleMat& first, const DoubleMat& second, int row, int col);
 
     /* Multiplies two matrices, the result is saved inside this matrix 
        Does not waste time or space */
@@ -356,7 +356,7 @@ public:
         assert(first.col_num() == this->row_num() && second.col_num() == this->col_num());
         for (int i = 0; i < first.col_num(); ++i) {
             for (int j = 0; j < second.col_num(); ++j) {
-                double res = 0.;
+                float res = 0.;
                 for (int k = 0; k < first.row_num(); ++k) {
                     res += first[k][i] * second[k][j];
                 }
@@ -374,7 +374,7 @@ public:
         assert(first.row_num() == this->row_num() && second.row_num() == this->col_num());
         for (int i = 0; i < first.row_num(); ++i) {
             for (int j = 0; j < second.row_num(); ++j) {
-                double res = 0.;
+                float res = 0.;
                 for (int k = 0; k < first.col_num(); ++k) {
                     res += first[i][k] * second[j][k];
                 }
@@ -393,10 +393,10 @@ public:
 };
 
 /* Computes dot product of row of first matrix and col of second matrix */
-double dot_row_col(const DoubleMat& first, const DoubleMat& second, int row, int col)
+float dot_row_col(const DoubleMat& first, const DoubleMat& second, int row, int col)
 {
     assert(first.col_num() == second.row_num());
-    double res = 0.;
+    float res = 0.;
     for (int i = 0; i < first.col_num(); ++i) {
         res += first[row][i] * second[i][col];
     }
@@ -404,10 +404,10 @@ double dot_row_col(const DoubleMat& first, const DoubleMat& second, int row, int
 }
 
 /* Computes dot product of row of matrix and vector */
-double dot_row_vec(const DoubleMat& mat, const DoubleVec& vec, int row)
+float dot_row_vec(const DoubleMat& mat, const DoubleVec& vec, int row)
 {
     assert(mat.col_num() == vec.size());
-    double res = 0.;
+    float res = 0.;
     for (int i = 0; i < mat.col_num(); ++i) {
         res += mat[row][i] * vec[i];
     }
@@ -415,10 +415,10 @@ double dot_row_vec(const DoubleMat& mat, const DoubleVec& vec, int row)
 }
 
 /* Computes dot product of vector and col of matrix */
-double dot_vec_col(const DoubleVec& vec, const DoubleMat& mat, int col)
+float dot_vec_col(const DoubleVec& vec, const DoubleMat& mat, int col)
 {
     assert(vec.size() == mat.row_num());
-    double res = 0.;
+    float res = 0.;
     for (int i = 0; i < vec.size(); ++i) {
         res += vec[i] * mat[i][col];
     }
@@ -433,17 +433,17 @@ DoubleMat operator-(DoubleMat x, const DoubleMat &y)
 { return x -= y; }
 
 /* multiplication of a matrix by a scalar - both sides */
-DoubleMat operator*(DoubleMat m, double scalar)
+DoubleMat operator*(DoubleMat m, float scalar)
 { return m *= scalar; }
 
-DoubleMat operator*(double scalar, DoubleMat m)
+DoubleMat operator*(float scalar, DoubleMat m)
 { return m *= scalar; }
 
 /* division of a matrix by a scalar - both sides */
-DoubleMat operator/(DoubleMat m, double scalar)
+DoubleMat operator/(DoubleMat m, float scalar)
 { return m /= scalar; }
 
-DoubleMat operator/(double scalar, DoubleMat m)
+DoubleMat operator/(float scalar, DoubleMat m)
 { return m /= scalar; }
 
 /* multiplication of a vector by a matrix - both sides 
@@ -507,7 +507,7 @@ DoubleMat sqrt_inside(DoubleMat mat)
 }
 
 /* adds given vector to all rows */
-DoubleMat add_scalar_to_all_items(DoubleMat mat, double d)
+DoubleMat add_scalar_to_all_items(DoubleMat mat, float d)
 {        
     for (int i = 0; i < mat.row_num(); ++i) {
         mat[i] = std::move(add_scalar_to_all_items(mat[i], d));
@@ -525,7 +525,7 @@ DoubleMat divide_by_items(DoubleMat first, const DoubleMat& second)
     return first;
 }
 
-/* printing a matrix, double printed to 4 decimal places */
+/* printing a matrix, float printed to 4 decimal places */
 inline void print_matrix(const DoubleMat &m)
 {
     std::cout << std::setprecision(4) << std::fixed;
